docs: Update to documentation
An added description to the what the link describes. This is a sample from Ewura from one of the courses and she is great at extracting data. Her code should help us navigate the web scrapping part of the project.
Her code features the following for us to take notice:
* Python library for extraction
*  Working with json